Unify counter dag types
Weird self refrence always existed via clock, it is just more obvious now and consists of one less layer.  It will be removed shortly.
