+++ internal/core/crdt/counter.go
@@ -11,19 +11,9 @@
 package crdt
 
 import (
-	"bytes"
-	"context"
-
-	"github.com/fxamacker/cbor/v2"
-	"github.com/sourcenetwork/corekv"
 	"golang.org/x/exp/constraints"
 
-	"github.com/sourcenetwork/defradb/client"
-	"github.com/sourcenetwork/defradb/datastore"
-	"github.com/sourcenetwork/defradb/errors"
 	"github.com/sourcenetwork/defradb/internal/core"
-	"github.com/sourcenetwork/defradb/internal/db/base"
-	"github.com/sourcenetwork/defradb/internal/keys"
 )
 
 type Incrementable interface {
@@ -70,143 +60,4 @@ func (delta *CounterDelta) GetPriority() uint64 {
 // SetPriority will set the priority for this delta.
 func (delta *CounterDelta) SetPriority(prio uint64) {
 	delta.Priority = prio
-}
-
-// Counter, is a simple CRDT type that allows increment/decrement
-// of an Int and Float data types that ensures convergence.
-type Counter struct {
-	store datastore.DSReaderWriter
-	key   keys.DataStoreKey
-
-	AllowDecrement bool
-	Kind           client.ScalarKind
-}
-
-var _ core.ReplicatedData = (*Counter)(nil)
-
-// NewCounter returns a new instance of the Counter with the given ID.
-func NewCounter(
-	store datastore.DSReaderWriter,
-	key keys.DataStoreKey,
-	allowDecrement bool,
-	kind client.ScalarKind,
-) Counter {
-	return Counter{
-		store:          store,
-		key:            key,
-		AllowDecrement: allowDecrement,
-		Kind:           kind,
-	}
-}
-
-// Merge implements ReplicatedData interface.
-// It merges two CounterRegisty by adding the values together.
-func (c Counter) Merge(ctx context.Context, delta core.Delta) error {
-	d, ok := delta.(*CounterDelta)
-	if !ok {
-		return ErrMismatchedMergeType
-	}
-
-	return c.incrementValue(ctx, d.Data, d.GetPriority())
-}
-
-func (c Counter) incrementValue(
-	ctx context.Context,
-	valueAsBytes []byte,
-	priority uint64,
-) error {
-	key := c.key.WithValueFlag()
-	marker, err := c.store.Get(ctx, c.key.ToPrimaryDataStoreKey().Bytes())
-	if err != nil && !errors.Is(err, corekv.ErrNotFound) {
-		return err
-	}
-	if bytes.Equal(marker, []byte{base.DeletedObjectMarker}) {
-		key = key.WithDeletedFlag()
-	}
-
-	var resultAsBytes []byte
-
-	switch c.Kind {
-	case client.FieldKind_NILLABLE_INT:
-		resultAsBytes, err = validateAndIncrement[int64](ctx, c.store, key, valueAsBytes, c.AllowDecrement)
-		if err != nil {
-			return err
-		}
-	case client.FieldKind_NILLABLE_FLOAT32:
-		resultAsBytes, err = validateAndIncrement[float32](ctx, c.store, key, valueAsBytes, c.AllowDecrement)
-		if err != nil {
-			return err
-		}
-	case client.FieldKind_NILLABLE_FLOAT64:
-		resultAsBytes, err = validateAndIncrement[float64](ctx, c.store, key, valueAsBytes, c.AllowDecrement)
-		if err != nil {
-			return err
-		}
-	default:
-		return NewErrUnsupportedCounterType(c.Kind)
-	}
-
-	err = c.store.Set(ctx, key.Bytes(), resultAsBytes)
-	if err != nil {
-		return NewErrFailedToStoreValue(err)
-	}
-
-	return setPriority(ctx, c.store, c.key, priority)
-}
-
-func (c Counter) CType() client.CType {
-	if c.AllowDecrement {
-		return client.PN_COUNTER
-	}
-	return client.P_COUNTER
-}
-
-func validateAndIncrement[T Incrementable](
-	ctx context.Context,
-	store datastore.DSReaderWriter,
-	key keys.DataStoreKey,
-	valueAsBytes []byte,
-	allowDecrement bool,
-) ([]byte, error) {
-	value, err := getNumericFromBytes[T](valueAsBytes)
-	if err != nil {
-		return nil, err
-	}
-
-	if !allowDecrement && value < 0 {
-		return nil, NewErrNegativeValue(value)
-	}
-
-	curValue, err := getCurrentValue[T](ctx, store, key)
-	if err != nil {
-		return nil, err
-	}
-
-	newValue := curValue + value
-	return cbor.Marshal(newValue)
-}
-
-func getCurrentValue[T Incrementable](
-	ctx context.Context,
-	store datastore.DSReaderWriter,
-	key keys.DataStoreKey,
-) (T, error) {
-	curValue, err := store.Get(ctx, key.Bytes())
-	if err != nil {
-		if errors.Is(err, corekv.ErrNotFound) {
-			return 0, nil
-		}
-		return 0, err
-	}
-
-	return getNumericFromBytes[T](curValue)
-}
-
-func getNumericFromBytes[T Incrementable](b []byte) (T, error) {
-	var val T
-	err := cbor.Unmarshal(b, &val)
-	if err != nil {
-		return val, err
-	}
-	return val, nil
 }
